refactor(core): Move data-table source control handling into the data-table module
Source control no longer imports data-table internals. A
SourceControlDataTableHandler interface plus a @Service() registry live in
source-control.ee; the data-table module implements the handler (export,
status listing, import/reconciliation, deletion) and registers it during
module init(). Queries that need TypeORM operators moved behind use-case
methods on DataTableRepository. When the data-table module is inactive,
source control skips the resource type instead of failing at query time.

Part of DEVP-646
