+++ packages/cli/src/modules/source-control.ee/source-control-export.service.ee.ts
@@ -19,12 +19,10 @@ import { UnexpectedError } from 'n8n-workflow';
 import { rm as fsRm, writeFile as fsWriteFile } from 'node:fs/promises';
 import path from 'path';
 
-import { DataTableRepository } from '@/modules/data-table/data-table.repository';
 import { formatWorkflow } from '@/workflows/workflow.formatter';
 
 import {
 	SOURCE_CONTROL_CREDENTIAL_EXPORT_FOLDER,
-	SOURCE_CONTROL_DATATABLES_EXPORT_FOLDER,
 	SOURCE_CONTROL_GIT_FOLDER,
 	SOURCE_CONTROL_PROJECT_EXPORT_FOLDER,
 	SOURCE_CONTROL_TAGS_EXPORT_FILE,
@@ -33,7 +31,6 @@ import {
 } from './constants';
 import {
 	getCredentialExportPath,
-	getDataTableExportPath,
 	getFoldersPath,
 	getProjectExportPath,
 	getVariablesPath,
@@ -46,7 +43,6 @@ import {
 import { SourceControlScopedService } from './source-control-scoped.service';
 import type { ExportResult } from './types/export-result';
 import type { ExportableCredential } from './types/exportable-credential';
-import type { DataTableResourceOwner, ExportableDataTable } from './types/exportable-data-table';
 import type { ExportableFolder } from './types/exportable-folders';
 import { ExportableProject } from './types/exportable-project';
 import { ExportableVariable } from './types/exportable-variable';
@@ -65,8 +61,6 @@ export class SourceControlExportService {
 
 	private credentialExportFolder: string;
 
-	private dataTableExportFolder: string;
-
 	constructor(
 		private readonly logger: Logger,
 		private readonly variablesService: VariablesService,
@@ -79,7 +73,6 @@ export class SourceControlExportService {
 		private readonly folderRepository: FolderRepository,
 		private readonly sourceControlScopedService: SourceControlScopedService,
 		instanceSettings: InstanceSettings,
-		private readonly dataTableRepository: DataTableRepository,
 	) {
 		this.gitFolder = path.join(instanceSettings.n8nFolder, SOURCE_CONTROL_GIT_FOLDER);
 		this.workflowExportFolder = path.join(this.gitFolder, SOURCE_CONTROL_WORKFLOW_EXPORT_FOLDER);
@@ -88,7 +81,6 @@ export class SourceControlExportService {
 			this.gitFolder,
 			SOURCE_CONTROL_CREDENTIAL_EXPORT_FOLDER,
 		);
-		this.dataTableExportFolder = path.join(this.gitFolder, SOURCE_CONTROL_DATATABLES_EXPORT_FOLDER);
 	}
 
 	getWorkflowPath(workflowId: string): string {
@@ -97,10 +89,6 @@ export class SourceControlExportService {
 
 	getCredentialsPath(credentialsId: string): string {
 		return getCredentialExportPath(credentialsId, this.credentialExportFolder);
-	}
-
-	getDataTablePath(dataTableId: string): string {
-		return getDataTableExportPath(dataTableId, this.dataTableExportFolder);
 	}
 
 	async deleteRepositoryFolder() {
@@ -249,129 +237,6 @@ export class SourceControlExportService {
 		} catch (error) {
 			this.logger.error('Failed to export variables to work folder', { error });
 			throw new UnexpectedError('Failed to export variables to work folder', {
-				cause: error,
-			});
-		}
-	}
-
-	async exportDataTablesToWorkFolder(
-		candidates: SourceControlledFile[],
-		context: SourceControlContext,
-	): Promise<ExportResult> {
-		try {
-			sourceControlFoldersExistCheck([this.gitFolder, this.dataTableExportFolder]);
-
-			if (candidates.length === 0) {
-				return {
-					count: 0,
-					folder: this.dataTableExportFolder,
-					files: [],
-				};
-			}
-
-			// Extract data table IDs from candidates
-			const candidateIds = candidates.map((candidate) => candidate.id);
-
-			// Fetch only the selected data tables
-			const dataTables = await this.dataTableRepository.find({
-				where: {
-					id: In(candidateIds),
-					...this.sourceControlScopedService.getDataTablesInAdminProjectsFromContextFilter(context),
-				},
-				relations: [
-					'columns',
-					'project',
-					'project.projectRelations',
-					'project.projectRelations.role',
-					'project.projectRelations.user',
-				],
-				select: {
-					id: true,
-					name: true,
-					projectId: true,
-					createdAt: true,
-					updatedAt: true,
-					columns: {
-						id: true,
-						name: true,
-						type: true,
-						index: true,
-					},
-					project: {
-						id: true,
-						name: true,
-						type: true,
-						projectRelations: {
-							userId: true,
-							role: {
-								slug: true,
-							},
-							user: {
-								email: true,
-							},
-						},
-					},
-				},
-			});
-
-			const exportedFiles: Array<{ id: string; name: string }> = [];
-
-			// Write each data table to its own file
-			for (const table of dataTables) {
-				let owner: DataTableResourceOwner | null = null;
-				if (table.project?.type === 'personal') {
-					const ownerRelation = table.project.projectRelations?.find(
-						(pr) => pr.role.slug === PROJECT_OWNER_ROLE_SLUG,
-					);
-					if (ownerRelation) {
-						owner = {
-							type: 'personal',
-							projectId: table.project.id,
-							projectName: table.project.name,
-							personalEmail: ownerRelation.user.email,
-						};
-					}
-				} else if (table.project?.type === 'team') {
-					owner = {
-						type: 'team',
-						teamId: table.project.id,
-						teamName: table.project.name,
-					};
-				}
-
-				const exportableDataTable: ExportableDataTable = {
-					id: table.id,
-					name: table.name,
-					columns: table.columns
-						.sort((a, b) => a.index - b.index)
-						.map((col) => ({
-							id: col.id,
-							name: col.name,
-							type: col.type,
-							index: col.index,
-						})),
-					ownedBy: owner,
-					createdAt: table.createdAt.toISOString(),
-					updatedAt: table.updatedAt.toISOString(),
-				};
-
-				const filePath = this.getDataTablePath(table.id);
-				await fsWriteFile(filePath, JSON.stringify(exportableDataTable, null, 2));
-
-				exportedFiles.push({
-					id: table.id,
-					name: filePath,
-				});
-			}
-
-			return {
-				count: dataTables.length,
-				folder: this.dataTableExportFolder,
-				files: exportedFiles,
-			};
-		} catch (error) {
-			this.logger.error('Failed to export data tables to work folder', { error });
-			throw new UnexpectedError('Failed to export data tables to work folder', {
 				cause: error,
 			});
 		}
